Move literal AST nodes to a dir "Literal"
